// src/app/shared/models/Customer.ts
export class Customer  {
    _id?: string;
    firstName: string;
    lastName: string;
    gender?: string;
    mobileNumber?: string;
    emailId?: string;
    profileImage?: string;
    dateOfBirth?: Date;
    dateOfAnniversary?: Date;
    groupCode?: [string];
    passport?;
    aadharCard?;
    gstId?;
    createdAt?: Date;
    updatedAt?: Date;
}